fix: move general settings into webiny-app-cms.
