+++ packages/webiny-api/src/entities/index.js
@@ -6,5 +6,4 @@ export { groups2entitiesFactory } from "webiny-api/entities/Groups2Entities.enti
 export { roleFactory } from "webiny-api/entities/Role.entity";
 export { roles2entitiesFactory } from "webiny-api/entities/Roles2Entities.entity";
 export { settingsFactory } from "webiny-api/entities/Settings.entity";
-export { generalSettingsFactory } from "webiny-api/entities/GeneralSettings.entity";
 export { userSettingsFactory } from "webiny-api/entities/UserSettings.entity";
