+++ packages/webiny-api/src/plugins/index.js
@@ -2,6 +2,5 @@
 import graphqlContextEntities from "./graphqlContextEntities";
 import entities from "./entities";
 import graphql from "./graphql";
-import generalSettings from "./settings/generalSettings";
 
-export default [...entities, graphql, graphqlContextEntities, generalSettings];
+export default [...entities, graphql, graphqlContextEntities];
