--- packages/webiny-app-admin/src/plugins/Settings/GeneralSettings/index.js
@@ -1,20 +1,0 @@
-// @flow
-import * as React from "react";
-import GeneralSettings from "./components/GeneralSettings";
-import type { SettingsPluginType } from "webiny-app-admin/types";
-
-export default ({
-    type: "settings",
-    name: "settings-general-settings",
-    settings: {
-        type: "other",
-        name: "General settings",
-        component: <GeneralSettings />,
-        route: {
-            name: "Settings.GeneralSettings",
-            path: "/general",
-            title: "General Settings",
-            group: undefined
-        }
-    }
-}: SettingsPluginType);
