--- packages/webiny-app-admin/src/plugins/Settings/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default as generalSettings } from "./GeneralSettings";
