+++ packages/webiny-app-cms/src/admin/plugins/settings/index.js
@@ -1,6 +1,7 @@
 // @flow
 import * as React from "react";
 import CmsSettings from "./components/CmsSettings";
+import GeneralSettings from "./components/generalSettings/GeneralSettings";
 import type { SettingsPluginType } from "webiny-app-admin/types";
 
 export default ([
@@ -15,6 +16,21 @@ export default ([
                 name: "Settings.Cms",
                 path: "/cms",
                 title: "CMS",
+                group: undefined
+            }
+        }
+    },
+    {
+        type: "settings",
+        name: "settings-general-settings",
+        settings: {
+            type: "other",
+            name: "General settings",
+            component: <GeneralSettings />,
+            route: {
+                name: "Settings.GeneralSettings",
+                path: "/general",
+                title: "General Settings",
                 group: undefined
             }
         }
